+++ engine/src/test/java/net/jqwik/engine/properties/RandomGeneratorsTests.java
@@ -213,7 +213,7 @@ class RandomGeneratorsTests {
 					range, 0,
 					RandomGenerators.defaultShrinkingTargetCalculator(range, 0)
 				)
-			).isInstanceOf(IllegalArgumentException.class);
+			).isInstanceOf(JqwikException.class);
 		}
 
 		@Example